add settings gui
