+++ gui_prompt_user.py
@@ -1,0 +1,5 @@
+import PySimpleGUI as gui
+
+def prompt_user(title, message):
+
+    gui.popup(title, message)
